Check and comments/docstring for sever.
